+++ hw_12/src/constants.js
@@ -1,0 +1,3 @@
+const baseApiUrl = "https://api.ratesapi.io/api/latest";
+
+export default baseApiUrl;
